+++ resteasy-spring-boot-starter-test/src/test/java/com/paypal/springboot/resteasy/CommonUseCasesIT.java
@@ -1,10 +1,11 @@
 package com.paypal.springboot.resteasy;
 
-import com.test.Application;
+import com.sample.app.Application;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 import org.springframework.boot.SpringApplication;
 import org.springframework.util.SocketUtils;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
@@ -13,18 +14,28 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
 /**
- * This is an integration test based on a simple sample application (see sample-app project)
+ * This is an integration test based on a simple sample application and
+ * very common use cases (see sample-app project)
  *
  * @author facarvalho
  */
-public class ResteasySpringBootStarterIT {
+public class CommonUseCasesIT {
 
     @BeforeClass
-    public void beforeClass() {
+    public void startingApplicationUp() {
         RestAssured.basePath = "sample-app";
         int port = SocketUtils.findAvailableTcpPort();
         RestAssured.port = port;
-        SpringApplication.run(Application.class, "--server.port=" + port).registerShutdownHook();
+
+        SpringApplication springApplication = new SpringApplication(Application.class);
+        springApplication.addListeners(new LogbackTestApplicationListener());
+        springApplication.run("--server.port=" + port).registerShutdownHook();
+    }
+
+    @AfterClass
+    public void shuttingDownApplication() {
+        Response response = given().basePath("/").post("/shutdown");
+        response.then().statusCode(200).body("message", equalTo("Shutting down, bye..."));
     }
 
     @Test