+++ resteasy-spring-boot-starter-test/src/test/java/com/test/multicontexttest/JaxrsApplication.java
@@ -1,0 +1,16 @@
+package com.test.multicontexttest;
+
+import org.springframework.stereotype.Component;
+
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
+
+/**
+ * JAX-RS application
+ *
+ * Created by facarvalho on 12/7/15.
+ */
+@Component
+@ApplicationPath("/sample-app/")
+public class JaxrsApplication extends Application {
+}