+++ resteasy-spring-boot-starter-test/src/test/java/com/test/multicontexttest/MultiContextTestApp.java
@@ -1,4 +1,4 @@
-package com.test;
+package com.test.multicontexttest;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -10,10 +10,10 @@ import org.springframework.boot.web.support.SpringBootServletInitializer;
  * Created by facarvalho on 12/7/15.
  */
 @SpringBootApplication
-public class Application extends SpringBootServletInitializer {
+public class MultiContextTestApp extends SpringBootServletInitializer {
 
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(MultiContextTestApp.class, args);
     }
 
 }