+++ sample-app/src/main/java/com/sample/app/Application.java
@@ -1,0 +1,19 @@
+package com.sample.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.support.SpringBootServletInitializer;
+
+/**
+ * SpringBoot entry point application
+ *
+ * Created by facarvalho on 12/7/15.
+ */
+@SpringBootApplication
+public class Application extends SpringBootServletInitializer {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+}